# Remove regexp use