Copy/paste from the most important commit:
    size: stop using regexp
    
    Using a regular expression brings in the whole regexp and regexp/syntax
    packages, which increase the resulting binary size by about 130K
    (Go 1.18.1, Linux/amd64).
    
    Besides, regular expressions are generally slow and incur some
    initialization overhead (to compile all global regexp.MustComplile
    variables). This, unlike the size difference, is not the main motivation
    for this commit, but it feels like it should have also been mentioned.
    
    A quick benchmark comparison shows a huge improvement (again, this is
    not why this is done, nevertheless it pleases the eye):
    
    name         old time/op    new time/op    delta
    ParseSize-4    10.6µs ± 3%     2.6µs ±29%  -75.10%  (p=0.002 n=6+6)
    
    name         old alloc/op   new alloc/op   delta
    ParseSize-4    3.26kB ± 0%    0.20kB ± 0%  -93.75%  (p=0.000 n=7+6)
    
    name         old allocs/op  new allocs/op  delta
    ParseSize-4      72.0 ± 0%      26.0 ± 0%  -63.89%  (p=0.000 n=7+6)
    
    Compatibility note: I chose to accept (rather than error out) input
    strings like ".4 Gb" as this is valid and makes sense, and the inability
    to process such input by the old implementation is more about the
    limitation of regex being used, rather than a decision to error out on
    those.

For the rest, please see commit messages.