@kolyshkin 2022-05-16
> also, can you drop the first commit for now (adding the go.mod)?

Done